feat!: add _fa_use_pro state to control pro sets
In type 0.12.0, a warning will be thrown if a font is not found. This
causes several warnings when using the pro sets. This commit adds a
state to control the pro sets. If the state is set to true, the pro sets
will be used; otherwise, the pro sets will not be used.

BREAKING CHANGE: The default value of the state is false. Thus,
the default behavior of the pro sets icons will be toufu.

Fix #16

Co-authored-by: Dennis Burgermeister <dennis@zumlernen.de>